requirements.txt
Flask-Mailman
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
Flask-Migrate
Flask-bcrypt
wtforms_sqlalchemy
plotly
pandas
